定时任务业务

+++ gas-pay/src/main/java/cn/demo/gas/pay/service/RechargeService.java
@@ -29,7 +29,7 @@ public class RechargeService {
     }
 
 
-    public int getRechargeCount(RechargeForm submit){
+    public long getRechargeCount(RechargeForm submit){
         QuerySubmit querySubmit = new QuerySubmit();
         submit.setParams(querySubmit);
